backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pillow==10.1.0
aiofiles==23.2.1
pydantic==2.5.0